test/e2e: provision runc as default CRI-O runtime.
crun behaves differently than runc for CRI exec. With runc, the
exec'ed process gets the same resources as the target container.
With crun it does not. We use exec to get the resources of test
containers (inject shell, read /proc/self/status). This wouldn't
give correct results with crun, so we must ensure we always test
with runc.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>