WiP: resmgr: handle better redundant UpdateContainer()s.
Try doing a better job of detecting and handling redundant
UpdateContainer requests. These are requests that update a
subset of the resources of a container, and to its current
values. Such requests are typically sent by an active CPU
Manager with static policy enabled in the kubelet.

If the update request is missing some values, fill those
in first from the containers current resources. Then try
to check if CRI-level resource estimates calculated from
these is redundant (identical to the original one during
creation). If it is short-circuit reply with the current
resource assignments. Otherwise pass it on to the policy.