balloons: allow balloon match by expressions.
Implement assigning containers to balloon instances by
container match expression defined in balloon definitions.
Match expressions are evaluated for containers which are
not explicitly assigned to any balloon by an annotation,
before namespace-based assignments. If an expression in a
balloon definition matches a container, that container is
assigned to a balloon instance of that definition.

For instance, given this balloon definition among the
configuration

...
  balloonTypes:
  - name: specialBalloon
    matchContainers:
      - key: pod/labels/app.kubernetes.io/component
        operator: Equals
        values: [ "SpecialComponent" ]
...

any container not explicitly annotated to some other balloon
will get assigned to instances of specialBalloon, if the pod
of the container has the label
  app.kubernetes.io/component=SpecialComponent

Similarly, given this balloon definition among the
configuration

...
  balloonTypes:
  - name: podBalloon
    matchContainers:
      - key: pod/name
        operator: In
        values: [ "pod1", "pod2", "pod3" ]
...

all containers of pods named pod1, pod2, and pod3 will be
assigned to instances of podBalloon.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>