WiP: cpuallocator: discover last-level cache groups.
Discover groups of CPUs sharing last-level cache. Ignore any
group which happens to be the same as a set of CPUs groupped
by some other topological criteria (package, die, etc. CPUs).
Assume that all CPUs within an LLC group are from the same
package and die, and share the same closest NUMA node. Verify
this assumption and abort via panic if it is not true.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>