resmgr: lifecycle overlap detection and workaround.
We recently discovered a problem with the generated stream of
container lifecycle events with some runtime versions. A side
effect of this is that we get Create/Stop events for multiple
container instances with seemingly overlapping lifecycle: the
latter instance get created before the former one is stopped.

When undetected, such a false overlap might cause overcommit
of resources, with both instances temporarily using the full
resource set of the container. As a workaround, we now track
containers also by fully qualified name ($namespace/pod/ctr)
and internally generate an event for releasing the resources
if the old instance whenever we notice that a creation event
would cause a duplicate instance for the same name.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>