topology-aware: rework building the topology pool tree.
Rework and simplify how we build the tree of pools based on HW
topology. Build the tree by walking the topology from socket to
dies to NUMA nodes closest to dies. Assign CPU and memory with
close CPUs in tandem as we build the tree. Assign CPUless NUMA
nodes at the end, to the same nodes a their closest NUMA nodes
with CPUs are assigned to.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>