config,crds: add RDT and block I/O class configuration.
Add RDT class and partition and block I/O class configuration.
If given and the corresponding controls are enabled, these are
used to configure RDT and block I/O classes using goresctrl.

For block I/O the configuration is practically identical to
the native goresctrl block I/O configuration apart from the
force flag which is passed to goresctrl/blockio/SetConfig()
as the force argument.

For RDT the configuration is almost identical to the native
goresctrl RDT configuration, but does not support the short-
hand convenience notation for unified caches which would be
problematic for a CR-based data type. IOW, one cannot simply
say

```
...
  l[23]Allocation:
    all: 75%
...
```

but instead needs to explicitly write this as

```
...
  l[23]Allocation:
    all:
      unified: 75%
...
```

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>